+++ examples/resources/iosxe_radius/resource.tf
@@ -1,8 +1,8 @@
 resource "iosxe_radius" "example" {
-  name                     = "radius_10.10.15.12"
-  radius_host_address_ipv4 = "10.10.15.12"
-  address_auth_port        = 1813
-  timeout                  = 4
-  retransmit               = 3
-  key_key                  = "123"
+  name                = "radius_10.10.15.12"
+  ipv4_address        = "10.10.15.12"
+  authentication_port = 1813
+  timeout             = 4
+  retransmit          = 3
+  key                 = "123"
 }